Updated impl of releases stats

+++ contracts/cw-cyber-gift/src/msg.rs
@@ -56,7 +56,7 @@ pub enum QueryMsg {
     IsClaimed { address: String },
     Claim { address: String },
     ReleaseState { address: String },
-    ReleaseStageState { stage: u32 },
+    ReleaseStageState { stage: u8 },
     AllReleaseStageStates{}
 }
